backend/requirements.txt
quart==0.20.0
python-dotenv==1.1.1
aiohttp==3.12.15
hypercorn==0.17.3
pandas==2.3.2
scikit-learn==1.7.2
xlrd==2.0.2
